#1818 Enhanced criterion hieararchy on client and server side. Enhnaced logic that determines what mnemonics are invisible and which not.
