+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/centre/api/crit/impl/AbstractMultiCriterionWidget.java
@@ -1,7 +1,5 @@
 package ua.com.fielden.platform.web.centre.api.crit.impl;
 
-import java.util.Map;
-
 import ua.com.fielden.platform.entity.AbstractEntity;
 import ua.com.fielden.platform.web.view.master.api.widgets.impl.AbstractWidget;
 
@@ -21,14 +19,5 @@ public abstract class AbstractMultiCriterionWidget extends AbstractCriterionWidg
      */
     public AbstractMultiCriterionWidget(final Class<? extends AbstractEntity<?>> root, final String widgetPath, final String propertyName, final AbstractWidget... editors) {
         super(root, widgetPath, propertyName, editors);
-    }
-
-    @Override
-    protected Map<String, Object> createCustomAttributes() {
-        final Map<String, Object> attrs = super.createCustomAttributes();
-        attrs.put("or-null", "{{propertyModel." + this.propertyName() + ".orNull}}");
-        attrs.put("not", "{{propertyModel." + this.propertyName() + ".not}}");
-        attrs.put("or-group", "{{propertyModel." + this.propertyName() + ".orGroup}}");
-        return attrs;
     }
 }
