+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/centre/api/crit/impl/StringCriterionWidget.java
@@ -18,7 +18,7 @@ public class StringCriterionWidget extends AbstractMultiCriterionWidget {
      * @param propertyName
      */
     public StringCriterionWidget(final Class<? extends AbstractEntity<?>> root, final Class<?> managedType, final String propertyName) {
-        super(root, "centre/criterion/multi/tg-multi-criterion", propertyName,
+        super(root, "centre/criterion/tg-criterion", propertyName,
                 new SinglelineTextWidget(
                         AbstractCriterionWidget.generateSingleTitleDesc(root, managedType, propertyName),
                         AbstractCriterionWidget.generateSingleName(root, managedType, propertyName)
