+++ platform-web-ui/src/main/web/ua/com/fielden/platform/web/centre/criterion/tg-criterion-config.js
@@ -1,6 +1,6 @@
 /**
- * The `tg-multi-criterion-config` contains just the DOM part of all non-single `tg-criterion`s -- 
- * 'Missing value' and 'Not' editors.
+ * The `tg-criterion-config` contains just the DOM part of all `tg-criterion`s -- 
+ * 'Missing value', 'Not' and 'OrGroup' editors.
  */
 import { Polymer } from '/resources/polymer/@polymer/polymer/lib/legacy/polymer-fn.js';
 import { html } from '/resources/polymer/@polymer/polymer/lib/utils/html-tag.js';
@@ -32,13 +32,13 @@ const template = html`
         </tg-flex-layout>
     </tg-accordion>
     <paper-checkbox checked="{{_orNull}}" hidden$="[[_excludeMissing]]">Missing</paper-checkbox>
-    <paper-checkbox checked="{{_not}}">Not</paper-checkbox>
+    <paper-checkbox checked="{{_not}}" hidden$="[[_excludeNot]]">Not</paper-checkbox>
 `;
 
 Polymer({
     _template: template,
 
-    is: 'tg-multi-criterion-config',
+    is: 'tg-criterion-config',
 
     properties: {
         _orNull: {
@@ -59,6 +59,9 @@ Polymer({
         },
 
         _excludeMissing: {
+            type: Boolean
+        },
+        _excludeNot: {
             type: Boolean
         },
         _excludeOrGroup: {
